requirements.txt
flet==0.21.2
pillow>=10.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
